+++ source/module_base/global_function.h
@@ -167,17 +167,17 @@ static inline void DCOPY( const T &a, T &b, const int &dim)
 }
 
 template<typename T>
-inline void COPYARRAY(const T* a, T* b, const long dim);
+inline void COPYARRAY(const T* a, T* b, const size_t dim);
 
 template<>
-inline void COPYARRAY(const std::complex<double>* a, std::complex<double>* b, const long dim)
+inline void COPYARRAY(const std::complex<double>* a, std::complex<double>* b, const size_t dim)
 {
     const int one = 1;
     zcopy_(&dim, a, &one, b, &one);
 }
 
 template<>
-inline void COPYARRAY(const double* a, double* b, const long dim)
+inline void COPYARRAY(const double* a, double* b, const size_t dim)
 {
     const int one = 1;
     dcopy_(&dim, a, &one, b, &one);
